Add configuration at global and instance levels
