+++ README.md
@@ -4,7 +4,7 @@
 
 This gem is framework agnostic and should work with any Rack based application including both [Rails][rails] and [Sinatra][sinatra].
 
-[![Build Status][shield-travis]][info-travis] [![Code Coverage][shield-coveralls]][info-coveralls] [![License][shield-license]][info-license]
+[![Gem Version][shield-gem]][info-gem] [![Build Status][shield-travis]][info-travis] [![Code Coverage][shield-coveralls]][info-coveralls] [![License][shield-license]][info-license]
 
 ## Contents
 <!-- START doctoc generated TOC please keep comment here to allow auto update -->
@@ -16,7 +16,11 @@ This gem is framework agnostic and should work with any Rack based application i
 - [Usage](#usage)
   - [Methods](#methods)
   - [Disabling checks](#disabling-checks)
-    - [Example](#example)
+    - [Examples](#examples)
+      - [Globally](#globally)
+      - [Instance level](#instance-level)
+        - [With a block](#with-a-block)
+        - [Calling `#configure`](#calling-configure)
   - [Handling errors](#handling-errors)
 - [Getting Started with Development](#getting-started-with-development)
   - [Running the tests](#running-the-tests)
@@ -86,6 +90,8 @@ You are also able to configure [AlexaRequestVerifier][alexa_request_verifier] to
 ### Disabling checks
 If you'd like to disable one (or more) of the checks performed by [AlexaRequestVerifier][alexa_request_verifier], you can do so by passing a #configure block. Each of the configuration attributes are Boolean values and are detailed below.
 
+It is possible to disable checks either globally, or for a specific instance. This is useful if you want to run multiple instances of the verifier within your application.
+
 Option | Default | Description
 ---|---|---
 `enabled` | `true` | Enables or disables AlexaRequestVerifier checks. This setting overrides all others i.e. setting `config.enabled = false` disables all checks even if you set others to true.
@@ -94,9 +100,10 @@ Option | Default | Description
 `verify_certificate` | `true` | Enables or disabled checks on whether the certificate is in date, or contains the SAN address we expect.
 `verify_signature` | `true` | Enables or disables checks to see if a request was actually signed by a certificate.
 
-#### Example
-The blow is an example of a 'complete' configure block.
+#### Examples
+The below is an example of a 'complete' configure block, setting attributes both globally and for an individual instance.
 
+##### Globally
 ```ruby
 AlexaRequestVerifier.configure do |config|
   config.enabled            = false # Disables all checks, even though we enable them individually below
@@ -105,6 +112,33 @@ AlexaRequestVerifier.configure do |config|
   config.verify_certificate = true
   config.verify_signature   = true
 end
+AlexaRequestVerifier.valid!(request)
+```
+
+##### Instance level
+###### With a block
+```ruby
+verifier = AlexaRequestVerifier::Verifier.new do |config|
+  config.enabled            = false
+  config.verify_uri         = true
+  config.verify_timeliness  = true
+  config.verify_certificate = true
+  config.verify_signature   = true
+end
+verifier.valid!(request)
+```
+
+###### Calling `#configure`
+```ruby
+verifier = AlexaRequestVerifier::Verifier.new
+verifier.configure do |config|
+  config.enabled            = false
+  config.verify_uri         = true
+  config.verify_timeliness  = true
+  config.verify_certificate = true
+  config.verify_signature   = true
+end
+verifier.valid!(request)
 ```
 
 
@@ -149,8 +183,8 @@ If you wish to submit a bug fix or feature, you can create a pull request and it
 ## License
 [AlexaRequestVerifier][alexa_request_verifier] is licensed under the [MIT][info-license].
 
-## Code of Conduct
 
+## Code of Conduct
 Everyone interacting in the AlexaRequestVerifier project’s codebases, issue trackers, chat rooms and mailing lists is expected to follow the [code of conduct][code_of_conduct].
 
 [alexa_request_verifier]: https://github.com/mattrayner/alexa_request_verifier
@@ -160,6 +194,9 @@ Everyone interacting in the AlexaRequestVerifier project’s codebases, issue tr
 [sinatra]:                http://sinatrarb.com
 [rspec]:                  http://rspec.info
 [code_of_conduct]:        https://github.com/mattrayner/alexa_request_verifier/blob/master/CODE_OF_CONDUCT.md
+
+[info-gem]:   https://rubygems.org/gems/alexa_request_verifier
+[shield-gem]: https://img.shields.io/gem/v/alexa_request_verifier.svg
 
 [info-travis]:   https://travis-ci.org/mattrayner/alexa_request_verifier
 [shield-travis]: https://img.shields.io/travis/mattrayner/alexa_request_verifier.svg
