+++ lib/alexa_request_verifier/verifier/certificate_verifier.rb
@@ -5,7 +5,7 @@ require 'openssl'
 require 'base64'
 
 module AlexaRequestVerifier
-  module Verifier
+  class Verifier
     # Given an OpenSSL certificate, validate it according to:
     # https://developer.amazon.com/docs/custom-skills/host-a-custom-skill-as-a-web-service.html#h2_verify_sig_cert
     #
