+++ spec/fixtures/vcr_cassettes/AlexaRequestVerifier/behaves_like_a_verifier_object/_configure/when_passing_a_block/that_disables_certificate_checking/returns_true_when_you_call_valid_with_an_invalid_certificate_in_our_store.yml
@@ -21,11 +21,11 @@ http_interactions:
       message: OK
     headers:
       X-Amz-Id-2:
-      - Xf7BECc0pzEDnkBsMrCBNWB+HjcmsvdjqZ7+0Aus/POfkM3s5b9c5+i3FhryWswovt0hMwAlFDU=
+      - erC1jkZ1ypqikxurxqwT/yFyi7mQz0jtNBcVf+S6P52nmwbWgOUIF4fYEXuGY9b7x5EBn2Lmzgg=
       X-Amz-Request-Id:
-      - 4263B8990A878618
+      - 14BB6C42AF1F35C9
       Date:
-      - Tue, 21 Nov 2017 11:02:38 GMT
+      - Wed, 22 Nov 2017 23:16:14 GMT
       Last-Modified:
       - Thu, 26 Oct 2017 21:36:16 GMT
       Etag:
@@ -106,5 +106,5 @@ http_interactions:
         Kvi4Os7X1g8RvmurFPW9QaAiY4nxug9vKWNmLT+sjHLF+8fk1A/yO0+MKcc=
         -----END CERTIFICATE-----
     http_version: 
-  recorded_at: Tue, 21 Nov 2017 00:00:00 GMT
+  recorded_at: Mon, 20 Nov 2017 08:45:05 GMT
 recorded_with: VCR 3.0.3
