+++ spec/fixtures/vcr_cassettes/AlexaRequestVerifier/behaves_like_a_verifier_object/_valid_/with_a_valid_request/returns_true.yml
@@ -21,11 +21,11 @@ http_interactions:
       message: OK
     headers:
       X-Amz-Id-2:
-      - dbIIhi+7Rx2XxqWy53jTidLJQ+T/SVswQ3mItj94+pEu0CSg+V7kQL8KpdcBHJbMT4g76o8eJTo=
+      - rEfcclvPe4JT+58Tk1A5NFFrQY116BtIvE75eZtVPphSB68kXCKO2hnhZhzCNXjnbKHYs1a51Ak=
       X-Amz-Request-Id:
-      - 2D036475B744C900
+      - D62AE2F61A51AAC5
       Date:
-      - Tue, 21 Nov 2017 23:14:35 GMT
+      - Wed, 22 Nov 2017 23:16:02 GMT
       Last-Modified:
       - Thu, 26 Oct 2017 21:36:16 GMT
       Etag:
