+++ spec/fixtures/vcr_cassettes/AlexaRequestVerifier/behaves_like_a_verifier_object/_valid_/with_an_invalid_request/removes_a_certificate_from_our_store_if_there_is_an_issue_verifying_a_request.yml
@@ -21,11 +21,11 @@ http_interactions:
       message: OK
     headers:
       X-Amz-Id-2:
-      - ogCNa06oL8bIREq+rQQIeWp2fU7pphC7k6uKV6GX+ZceGGmH4M/NY52Kz4oAeDH3KrPb5O4JVr4=
+      - 7c/7hCvH/JcQu1NhXFPV08CYjZoTm2mUbGEHjh/5qptFezGiHh5DsAN3tWIrn24aZtx9eqRS8zM=
       X-Amz-Request-Id:
-      - 223A825EE67154E2
+      - F7B6E085F7841B80
       Date:
-      - Tue, 21 Nov 2017 10:52:13 GMT
+      - Wed, 22 Nov 2017 23:15:59 GMT
       Last-Modified:
       - Thu, 26 Oct 2017 21:36:16 GMT
       Etag:
@@ -106,5 +106,5 @@ http_interactions:
         Kvi4Os7X1g8RvmurFPW9QaAiY4nxug9vKWNmLT+sjHLF+8fk1A/yO0+MKcc=
         -----END CERTIFICATE-----
     http_version: 
-  recorded_at: Tue, 21 Nov 2017 00:00:00 GMT
+  recorded_at: Mon, 20 Nov 2017 08:45:05 GMT
 recorded_with: VCR 3.0.3
