+++ spec/fixtures/vcr_cassettes/AlexaRequestVerifier_Verifier/behaves_like_a_verifier_object/_valid_/with_a_valid_timely_request/returns_true.yml
@@ -21,11 +21,11 @@ http_interactions:
       message: OK
     headers:
       X-Amz-Id-2:
-      - i4TXDPo51VqO0xV7tMDwDeOjg8Kr4AjNX+v37J9HOGfZUX8AU7yUb9YAw5GmRJhOG8qBJqqGOWU=
+      - FL4sa/H8mzRi0We5EHY/QnT6rM8eXwZY21IJj8bSUjF7yjR3k6yk3Y3WZJxtYo/4cwdpr+7YQx4=
       X-Amz-Request-Id:
-      - 7E1F5F29B4087598
+      - 4D3C8E05DF02F35F
       Date:
-      - Tue, 21 Nov 2017 11:04:47 GMT
+      - Wed, 22 Nov 2017 23:15:24 GMT
       Last-Modified:
       - Thu, 26 Oct 2017 21:36:16 GMT
       Etag:
